requirements.txt
# Project requirements

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23

# Spatial libraries
shapely==2.0.2
pyproj==3.6.1
rtree==1.1.0
geohash2==1.1.0

# Caching
redis==5.0.1
aioredis==2.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
httpx==0.27.0

# Data generation
faker==19.13.0
numpy==1.25.2

# Monitoring
prometheus-client==0.19.0
